# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a professional ecommerce server for Eraya.",
        "title": "Eraya Ecommerce API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "eraya-backend.onrender.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/chat/conversation/{id}": {
            "delete": {
                "description": "Permantently delete a conversation and all its messages. Admin only.",
                "tags": [
                    "chat"
                ],
                "summary": "Delete a conversation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Conversation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/chat/conversation/{withID}": {
            "get": {
                "description": "Fetch all messages between the current user and another user.",
                "tags": [
                    "chat"
                ],
                "summary": "Get messages in a conversation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Other User ID",
                        "name": "withID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Message"
                            }
                        }
                    }
                }
            }
        },
...
        "/chat/ws": {
            "get": {
                "description": "Establish a WebSocket connection for real-time messaging. Requires token in query param.",
                "tags": [
                    "chat"
                ],
                "summary": "Connect to real-time chat",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Chat with User ID",
                        "name": "with",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "domain.Message": {
            "type": "object",
            "properties": {
                "attachment_url": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_read": {
                    "type": "boolean"
                },
                "message_text": {
                    "type": "string"
                },
                "online": {
                    "type": "boolean"
                },
                "order_id": {
                    "type": "integer"
                },
                "receiver_id": {
                    "type": "integer"
                },
                "reply_to_id": {
                    "type": "integer"
                },
                "reply_to_text": {
                    "type": "string"
                },
                "sender_avatar": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "integer"
                },
                "sender_name": {
                    "type": "string"
                },
                "temp_id": {
                    "type": "string"
                },
                "type": {
                    "description": "\"message\", \"presence\", \"update\", \"delete\"",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}